+++ integration/lql_create_test.go
@@ -71,7 +71,7 @@ func TestQueryCreateFile(t *testing.T) {
 
 	// create
 	out, stderr, exitcode := LaceworkCLIWithTOMLConfig("query", "create", "-f", file.Name())
-	assert.Contains(t, out.String(), fmt.Sprintf("query (%s) created successfully.", queryID))
+	assert.Contains(t, out.String(), fmt.Sprintf("Query (%s) created successfully.", queryID))
 	assert.Empty(t, stderr.String(), "STDERR should be empty")
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 
@@ -88,7 +88,7 @@ func TestQueryCreateFile(t *testing.T) {
 
 	// delete
 	out, stderr, exitcode = LaceworkCLIWithTOMLConfig("query", "delete", queryID)
-	assert.Contains(t, out.String(), fmt.Sprintf("query (%s) deleted successfully.", queryID))
+	assert.Contains(t, out.String(), fmt.Sprintf("Query (%s) deleted successfully.", queryID))
 	assert.Empty(t, stderr.String(), "STDERR should be empty")
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 }