+++ integration/lql_test.go
@@ -116,7 +116,7 @@ func TestQueryRunEditor(t *testing.T) {
 	out, err, exitcode = LaceworkCLIWithTOMLConfig("query", "run", "--validate_only")
 	assert.Contains(t, out.String(), "Type a query to validate")
 	assert.Contains(t, out.String(), "[Enter to launch editor]")
-	assert.Contains(t, err.String(), "ERROR unable to run query: EOF")
+	assert.Contains(t, err.String(), "ERROR unable to validate query: EOF")
 	assert.Equal(t, 1, exitcode, "EXITCODE is not the expected one")
 }
 
@@ -136,9 +136,8 @@ func TestQueryRunID(t *testing.T) {
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 
 	// validate_only
-	out, err, exitcode = LaceworkCLIWithTOMLConfig("query", "run", queryID, "--validate_only")
-	assert.Contains(t, out.String(), "query validated successfully.")
-	assert.Empty(t, err.String(), "STDERR should be empty")
+	_, err, exitcode = LaceworkCLIWithTOMLConfig("query", "run", queryID, "--validate_only")
+	assert.Contains(t, err.String(), "ERROR flag --validate_only unavailable when specifying query_id argument")
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 }
 
@@ -221,7 +220,7 @@ func TestQueryRunFile(t *testing.T) {
 	// validate_only
 	out, stderr, exitcode = LaceworkCLIWithTOMLConfig(
 		"query", "run", "-f", file.Name(), "--validate_only")
-	assert.Contains(t, out.String(), "query validated successfully.")
+	assert.Contains(t, out.String(), "Query validated successfully.")
 	assert.Empty(t, stderr.String(), "STDERR should be empty")
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 }
@@ -240,7 +239,7 @@ func TestQueryRunURL(t *testing.T) {
 	// validate_only
 	out, err, exitcode = LaceworkCLIWithTOMLConfig(
 		"query", "run", "-u", queryURL, "--validate_only")
-	assert.Contains(t, out.String(), "query validated successfully.")
+	assert.Contains(t, out.String(), "Query validated successfully.")
 	assert.Empty(t, err.String(), "STDERR should be empty")
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 }