+++ RELEASE_NOTES.md
@@ -1,9 +1,17 @@
 # Release Notes
-Another day, another release. These are the release notes for the version `v0.9.1`.
+Another day, another release. These are the release notes for the version `v0.11.0`.
 
+## Features
+* feat(cli): manage EmailUser alert channels (Salim Afiune Maya)([4624dfb](https://github.com/lacework/go-sdk/commit/4624dfb6524f653f98eae7c87357d99c247e6f6c))
+* feat(api): EmailUser Alert Channel (v2) (#492) (Salim Afiune)([45cc8f4](https://github.com/lacework/go-sdk/commit/45cc8f4eef3e0eacac66f4cb30be47b20fec9a6b))
+* feat(cli): apiv2 support for query run (hazedav)([6d8c766](https://github.com/lacework/go-sdk/commit/6d8c766b3838cbfe45ce874c3741820aeafdd5bc))
 ## Bug Fixes
-* fix(api): Parse LastEvaluationTime correctly in hostVulnListHostsResponse (#446) (Darren)([7918c4c](https://github.com/lacework/go-sdk/commit/7918c4c2cf816cad193f07a35150ce8de91481e0))
-* fix(api): Enable FirstSeenTime field in hostVulnListHostsResponse (#445) (Darren)([195d44a](https://github.com/lacework/go-sdk/commit/195d44a98d019b293edaf8643c8aa876b996b945))
+* fix(cli): integration show align left (Salim Afiune Maya)([df17625](https://github.com/lacework/go-sdk/commit/df176256fe0f8cd49bc95298debe3001c539933d))
+* fix(cli): integration show single object (Salim Afiune Maya)([b26a3d0](https://github.com/lacework/go-sdk/commit/b26a3d07f062ee39be6ba745b052de49ba4ccea5))
+* fix(cli): evalFrequency optional for policyCreate (hazedav)([03d527e](https://github.com/lacework/go-sdk/commit/03d527edb8e02229625176ed30548c50a43c0e92))
+## Other Changes
+* chore(cli): apply lint recommendations (Salim Afiune Maya)([931b2db](https://github.com/lacework/go-sdk/commit/931b2db86daad2cfa11f97aa6dd3766ee06ff162))
+* test(cli): re-enable on-demand ctr scans (#490) (Salim Afiune)([ddddbc6](https://github.com/lacework/go-sdk/commit/ddddbc66aad78c4f69de7bb0104c4c5e80a058b9))
 
 ## Docker Images
 * `docker pull lacework/lacework-cli:latest`